+++ cis-compute-storage/volume-groups.tf
@@ -2,13 +2,15 @@
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
 locals {
-  volume_groups_to_replicate = { for k, v in(var.storage_configuration != null ? (var.storage_configuration["volume_groups"] != null ? var.storage_configuration["volume_groups"] : {}) : {}) : k => v if v.replication != null }
+  volume_groups = try(coalesce(var.storage_configuration.volume_groups, {}), {})
 
-  volume_groups_with_backup_policies = { for k, v in(var.storage_configuration != null ? (var.storage_configuration["volume_groups"] != null ? var.storage_configuration["volume_groups"] : {}) : {}) : k => v if v.backup.enable_backup == true && v.backup_policy != null }
+  volume_groups_to_replicate         = { for k, v in local.volume_groups : k => v if v.replication != null }
+  volume_groups_with_backups         = { for k, v in local.volume_groups : k => v if try(v.backup.enable_backup, false) == true }
+  volume_groups_with_backup_policies = { for k, v in local.volume_groups : k => v if v.backup_policy != null }
 }
 
 data "oci_identity_availability_domains" "vg_ads" {
-  for_each       = var.storage_configuration != null ? (var.storage_configuration["volume_groups"] != null ? var.storage_configuration["volume_groups"] : {}) : {}
+  for_each       = local.volume_groups
   compartment_id = each.value.compartment_id != null ? (length(regexall("^ocid1.*$", each.value.compartment_id)) > 0 ? each.value.compartment_id : var.compartments_dependency[each.value.compartment_id].id) : (length(regexall("^ocid1.*$", var.storage_configuration.default_compartment_id)) > 0 ? var.storage_configuration.default_compartment_id : var.compartments_dependency[var.storage_configuration.default_compartment_id].id)
 }
 
@@ -19,7 +21,7 @@ data "oci_identity_availability_domains" "vg_ads_replicas" {
 }
 
 resource "oci_core_volume_group" "these" {
-  for_each            = var.storage_configuration != null ? (var.storage_configuration["volume_groups"] != null ? var.storage_configuration["volume_groups"] : {}) : {}
+  for_each            = local.volume_groups
   availability_domain = data.oci_identity_availability_domains.vg_ads[each.key].availability_domains[each.value.availability_domain - 1].name
   compartment_id      = each.value.compartment_id != null ? (length(regexall("^ocid1.*$", each.value.compartment_id)) > 0 ? each.value.compartment_id : var.compartments_dependency[each.value.compartment_id].id) : (length(regexall("^ocid1.*$", var.storage_configuration.default_compartment_id)) > 0 ? var.storage_configuration.default_compartment_id : var.compartments_dependency[var.storage_configuration.default_compartment_id].id)
   source_details {
@@ -48,9 +50,7 @@ resource "oci_core_volume_group" "these" {
 }
 
 resource "oci_core_volume_group_backup" "these" {
-  for_each = { for k, v in try(var.storage_configuration.volume_groups, {}) : k => v
-    if v.backup.enable_backup == true ## enable_backup must be true to create a volume group backup
-  }
+  for_each        = local.volume_groups_with_backups
   volume_group_id = oci_core_volume_group.these[each.key].id
   type            = each.value.backup.type != null ? each.value.backup.type : "INCREMENTAL"
   display_name    = each.value.display_name != null ? "${each.value.display_name}-backup" : "volume-group-backup"